chore: refactor api libs
